requirements.txt
notebook
torch

numpy
matplotlib
seaborn

pillow
imageio
scipy
scikit-image

opencv

sklearn
sphinx